This commit has the null error fixed, but if u run the program and play strategi u will see 3 printstatements called player.
Currently only 1 player is intialized, and the print is supposed to show that id, but as u can tell it shows 2 diffent ID's. This needs to be fixed first before we can continue.

Also player2 instance is currently turned off until we have fixed this.

lastely, onbuttongridclicked has to be rewriten in its entirity, and please read all the changes i made to this document so u dont create stuff that isnt nessecary